Added more methods for validating and loading string datasets. (#2)
Of particular interest is the comprehensive checks for NULL pointers in
variable length strings; all functions dealing with vlen strings now throw an
error if a NULL is encountered. Additionally, we added the validate_*_string
functions to check whether a dataset/attribute has a vlen string type with NULL
pointers, without spending the time/memory to load it into a std::string.

Variable length string pointers are now freed via the usual RAII policy, with a
new stub class (VariableStringCleaner) responsible for handling freeing in its
destructor. This ensures that we don't have memory leaks when an error is
thrown before we manually free, e.g., because of a NULL pointer.

Finally, we added the load_scalar_string_dataset and load_1d_string_dataset
functions, to enable quick loading of string data. The latter is just a
wrapper around the Stream1dStringDataset loop for convenience.