Added back some helpful utilities by popular demand.
I ended up having to implement each of these separately in the
downstream libraries, so it's better to just have it here. Technically
they mix a few validation concerns together, so we'll just plonk them in
a 'miscellaneous' file to indicate that support for different
combinations of validation steps is ad-hoc, added on an as-needed basis.