Converted open() into a template function so that constexpr works.
Now it accepts all manner of path types, from C-strings to std::strings
to filesystem::paths. Besides being more flexible and user-friendly,
this is necessary for constexpr to actually skip the type checks based
on the templated Path_ type; otherwise we get compilation errors on
Windows due to the underlying wchar_t type.