Force IEEE754 compliance during the float limit checks.
This restricts the set of allowed float datatypes to the IEEE_* predefined
types. It ensures that we have the correct split of bits between the exponent
and significant; otherwise, a custom float types could be used that provides
the necessary total bits of precision but doesn't actually store floats
faithfully, e.g., if everything was allocated to the exponent. This also
ensures that IEEE special values can be represented.

In practice, it seems that everything was already using an IEEE float datatype
anyway (or, for specific CPUs, an alias to the IEEE float) so this change
shouldn't have any major effect.